requirements.txt
Django==3.0.1
bpython==0.18
djangorestframework==3.11.0
Markdown==3.1.1
django-filter==2.2.0
lxml==4.4.2
django-crispy-forms==1.8.1
dj-database-url==0.5.0
psycopg2==2.8.4
